pyproject.toml
[project]
name = "claim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "anyio>=4.5",
    "httpx>=0.27.1",
    "pydantic>=2.11.0",
	"fastapi>=0.128.0",
	"uvicorn>=0.31.1",
	"mcp>=1.25.0",
	"fhir.resources>=8.1.0"
]
